hardcoded complex64 fixed

+++ experiments/kernel_learning/fsdn_vs_baselines.py
@@ -58,8 +58,7 @@ def make_hmk():
     eta = torch.tensor([[1.0]])
     frequencies = torch.cat([eta, -eta], dim=0)
     B = torch.tensor(
-        [[2.0 + 0.0j, 0.0 + 0.5j], [0.0 - 0.5j, 2.0 + 0.0j]], dtype=torch.complex64
-    )
+        [[2.0 + 0.0j, 0.0 + 0.5j], [0.0 - 0.5j, 2.0 + 0.0j]])
     sigma1 = torch.eye(d) * (1.0 / (math.pi**2))
     sigma2 = torch.eye(d) * (1.0 / (2.0 * math.pi) ** 2)
     return HarmonizableMixtureKernel(
